Fix : Ripple TS configured
