+++ packages/component/src/button/Ripple.tsx
@@ -1,0 +1,376 @@
+import PropTypes from 'prop-types';
+import React, { PureComponent } from 'react';
+import {
+	View,
+	Animated,
+	Easing,
+	Platform,
+	TouchableWithoutFeedback,
+	I18nManager,
+	StyleSheet,
+} from 'react-native';
+
+const radius = 10;
+
+type RippleProps = {
+	rippleColor?: string;
+	rippleOpacity?: number;
+	rippleDuration?: number;
+	rippleSize?: number;
+	rippleContainerBorderRadius?: number;
+	rippleCentered?: boolean;
+	rippleSequential?: boolean;
+	rippleFades?: boolean;
+	disabled?: boolean;
+	onPress?: Function;
+	onPressIn?: Function;
+	onPressOut?: Function;
+	onLongPress?: Function;
+	onRippleAnimation?: Function;
+	onLayout?: Function;
+	testID?: any;
+	nativeID?: any;
+	accessible?: any;
+	accessibilityHint?: any;
+	accessibilityLabel?: string;
+	delayLongPress?: Function;
+	delayPressIn?: Function;
+	delayPressOut?: Function;
+	hitSlop?: any;
+	pressRetentionOffset?: any;
+};
+type RippleState = {
+	width: number;
+	height: number;
+	ripples: any[];
+};
+export default class Ripple extends PureComponent<
+	RippleProps,
+	RippleState
+> {
+	static defaultProps = {
+		...TouchableWithoutFeedback.defaultProps,
+
+		rippleColor: 'rgb(0, 0, 0)',
+		rippleOpacity: 0.3,
+		rippleDuration: 400,
+		rippleSize: 0,
+		rippleContainerBorderRadius: 0,
+		rippleCentered: false,
+		rippleSequential: false,
+		rippleFades: true,
+		disabled: false,
+
+		onRippleAnimation: (animation, callback) =>
+			animation.start(callback),
+	};
+
+	static propTypes = {
+		...Animated.View.propTypes,
+		...TouchableWithoutFeedback.propTypes,
+
+		rippleColor: PropTypes.string,
+		rippleOpacity: PropTypes.number,
+		rippleDuration: PropTypes.number,
+		rippleSize: PropTypes.number,
+		rippleContainerBorderRadius: PropTypes.number,
+		rippleCentered: PropTypes.bool,
+		rippleSequential: PropTypes.bool,
+		rippleFades: PropTypes.bool,
+		disabled: PropTypes.bool,
+
+		onRippleAnimation: PropTypes.func,
+	};
+
+	constructor(props) {
+		super(props);
+
+		this.onLayout = this.onLayout.bind(this);
+		this.onPress = this.onPress.bind(this);
+		this.onPressIn = this.onPressIn.bind(this);
+		this.onPressOut = this.onPressOut.bind(this);
+		this.onLongPress = this.onLongPress.bind(this);
+		this.onAnimationEnd =
+			this.onAnimationEnd.bind(this);
+
+		this.renderRipple =
+			this.renderRipple.bind(this);
+
+		//@ts-ignore
+		this.unique = 0;
+		//@ts-ignore
+		this.mounted = false;
+		//@ts-ignore
+		this.state = {
+			width: 0,
+			height: 0,
+			ripples: [],
+		};
+	}
+
+	componentDidMount() {
+		//@ts-ignore
+		this.mounted = true;
+	}
+
+	componentWillUnmount() {
+		//@ts-ignore
+		this.mounted = false;
+	}
+
+	onLayout(event) {
+		let { width, height } =
+			event.nativeEvent.layout;
+		let { onLayout } = this.props;
+
+		if ('function' === typeof onLayout) {
+			onLayout(event);
+		}
+
+		this.setState({ width, height });
+	}
+
+	onPress(event) {
+		let { ripples } = this.state;
+		let { onPress, rippleSequential } = this.props;
+
+		if (!rippleSequential || !ripples.length) {
+			if ('function' === typeof onPress) {
+				requestAnimationFrame(() => onPress(event));
+			}
+
+			this.startRipple(event);
+		}
+	}
+
+	onLongPress(event) {
+		let { onLongPress } = this.props;
+
+		if ('function' === typeof onLongPress) {
+			requestAnimationFrame(() =>
+				onLongPress(event),
+			);
+		}
+
+		this.startRipple(event);
+	}
+
+	onPressIn(event) {
+		let { onPressIn } = this.props;
+
+		if ('function' === typeof onPressIn) {
+			onPressIn(event);
+		}
+	}
+
+	onPressOut(event) {
+		let { onPressOut } = this.props;
+
+		if ('function' === typeof onPressOut) {
+			onPressOut(event);
+		}
+	}
+
+	onAnimationEnd() {
+		//@ts-ignore
+		if (this.mounted) {
+			this.setState(({ ripples }) => ({
+				ripples: ripples.slice(1),
+			}));
+		}
+	}
+
+	startRipple(event) {
+		let { width, height } = this.state;
+		let {
+			rippleDuration,
+			rippleCentered,
+			rippleSize,
+			onRippleAnimation,
+		} = this.props;
+
+		let w2 = 0.5 * width;
+		let h2 = 0.5 * height;
+
+		let { locationX, locationY } = rippleCentered
+			? { locationX: w2, locationY: h2 }
+			: event.nativeEvent;
+
+		let offsetX = Math.abs(w2 - locationX);
+		let offsetY = Math.abs(h2 - locationY);
+
+		let R =
+			rippleSize > 0
+				? 0.5 * rippleSize
+				: Math.sqrt(
+						Math.pow(w2 + offsetX, 2) +
+							Math.pow(h2 + offsetY, 2),
+				  );
+		let ripple = {
+			//@ts-ignore
+			unique: this.unique++,
+			progress: new Animated.Value(0),
+			locationX,
+			locationY,
+			R,
+		};
+
+		let animation = Animated.timing(
+			ripple.progress,
+			{
+				toValue: 1,
+				easing: Easing.out(Easing.ease),
+				duration: rippleDuration,
+				useNativeDriver: true,
+			},
+		);
+
+		onRippleAnimation(
+			animation,
+			this.onAnimationEnd,
+		);
+
+		this.setState(({ ripples }) => ({
+			ripples: ripples.concat(ripple),
+		}));
+	}
+
+	renderRipple({
+		unique,
+		progress,
+		locationX,
+		locationY,
+		R,
+	}) {
+		let {
+			rippleColor,
+			rippleOpacity,
+			rippleFades,
+		} = this.props;
+
+		let rippleStyle = {
+			top: locationY - radius,
+			[I18nManager.isRTL ? 'right' : 'left']:
+				locationX - radius,
+			backgroundColor: rippleColor,
+
+			transform: [
+				{
+					scale: progress.interpolate({
+						inputRange: [0, 1],
+						outputRange: [0.5 / radius, R / radius],
+					}),
+				},
+			],
+
+			opacity: rippleFades
+				? progress.interpolate({
+						inputRange: [0, 1],
+						outputRange: [rippleOpacity, 0],
+				  })
+				: rippleOpacity,
+		};
+
+		return (
+			<Animated.View
+				style={[styles.ripple, rippleStyle]}
+				key={unique}
+			/>
+		);
+	}
+
+	render() {
+		let { ripples } = this.state;
+		let {
+			delayLongPress,
+			delayPressIn,
+			delayPressOut,
+			disabled,
+			hitSlop,
+			pressRetentionOffset,
+			children,
+			rippleContainerBorderRadius,
+			testID,
+			nativeID,
+			accessible,
+			accessibilityHint,
+			accessibilityLabel,
+
+			onPress,
+			onLongPress,
+			onLayout,
+			onRippleAnimation,
+
+			rippleColor,
+			rippleOpacity,
+			rippleDuration,
+			rippleSize,
+			rippleCentered,
+			rippleSequential,
+			rippleFades,
+
+			...props
+		} = this.props;
+
+		let touchableProps = {
+			delayLongPress,
+			delayPressIn,
+			delayPressOut,
+			disabled,
+			hitSlop,
+			pressRetentionOffset,
+			testID,
+			accessible,
+			accessibilityHint,
+			accessibilityLabel,
+			onLayout: this.onLayout,
+			onPress: this.onPress,
+			onPressIn: this.onPressIn,
+			onPressOut: this.onPressOut,
+			onLongPress: onLongPress
+				? this.onLongPress
+				: undefined,
+
+			...('web' !== Platform.OS
+				? { nativeID }
+				: null),
+		};
+
+		let containerStyle = {
+			borderRadius: rippleContainerBorderRadius,
+		};
+
+		return (
+			<TouchableWithoutFeedback {...touchableProps}>
+				<Animated.View
+					{...props}
+					pointerEvents='box-only'
+				>
+					{children}
+					<View
+						style={[styles.container, containerStyle]}
+					>
+						{ripples.map(this.renderRipple)}
+					</View>
+				</Animated.View>
+			</TouchableWithoutFeedback>
+		);
+	}
+}
+
+const styles = StyleSheet.create({
+	container: {
+		...StyleSheet.absoluteFillObject,
+
+		backgroundColor: 'transparent',
+		overflow: 'hidden',
+	},
+
+	ripple: {
+		width: radius * 2,
+		height: radius * 2,
+		borderRadius: radius,
+		overflow: 'hidden',
+		position: 'absolute',
+	},
+});
